# docs/openapi.json
{
  "components": {
    "schemas": {
      "AnalyticsResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyClickCount"
            },
            "title": "Daily",
            "type": "array"
          },
          "top_referrers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Referrers",
            "type": "array"
          },
          "total_clicks": {
            "title": "Total Clicks",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "total_clicks",
          "daily",
          "top_referrers"
        ],
        "title": "AnalyticsResponse",
        "type": "object"
      },
      "ApprovalRequest": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "reviewer": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "approved",
          "reviewer"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "CreateURLRequest": {
        "properties": {
          "custom_alias": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 4,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Alias"
          },
          "expires_in_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateURLRequest",
        "type": "object"
      },
      "DailyClickCount": {
        "properties": {
          "clicks": {
            "title": "Clicks",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "clicks"
        ],
        "title": "DailyClickCount",
        "type": "object"
      },
      "EngineeringRunRequest": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "requirement": {
            "maxLength": 10000,
            "minLength": 10,
            "title": "Requirement",
            "type": "string"
          },
          "scenario_type": {
            "default": "greenfield",
            "enum": [
              "greenfield",
              "brownfield",
              "ambiguous"
            ],
            "title": "Scenario Type",
            "type": "string"
          }
        },
        "required": [
          "requirement"
        ],
        "title": "EngineeringRunRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ReplanRequest": {
        "properties": {
          "changed_requirement": {
            "maxLength": 10000,
            "minLength": 10,
            "title": "Changed Requirement",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "changed_requirement",
          "reason"
        ],
        "title": "ReplanRequest",
        "type": "object"
      },
      "URLResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disabled": {
            "title": "Disabled",
            "type": "boolean"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "short_url": {
            "title": "Short Url",
            "type": "string"
          },
          "target_url": {
            "title": "Target Url",
            "type": "string"
          }
        },
        "required": [
          "code",
          "target_url",
          "short_url",
          "created_at",
          "expires_at",
          "disabled"
        ],
        "title": "URLResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "URL shortener plus governed agentic SDLC orchestration prototype.",
    "title": "Agentic URL Shortener",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/engineering/metrics": {
      "get": {
        "operationId": "engineering_metrics_api_v1_engineering_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Engineering Metrics Api V1 Engineering Metrics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Engineering Metrics"
      }
    },
    "/api/v1/engineering/runs": {
      "post": {
        "operationId": "start_engineering_run_api_v1_engineering_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineeringRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Engineering Run Api V1 Engineering Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Engineering Run"
      }
    },
    "/api/v1/engineering/runs/{run_id}": {
      "get": {
        "operationId": "get_engineering_run_api_v1_engineering_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Engineering Run Api V1 Engineering Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Engineering Run"
      }
    },
    "/api/v1/engineering/runs/{run_id}/approvals/{node_id}": {
      "post": {
        "operationId": "approve_engineering_gate_api_v1_engineering_runs__run_id__approvals__node_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Engineering Gate Api V1 Engineering Runs  Run Id  Approvals  Node Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Engineering Gate"
      }
    },
    "/api/v1/engineering/runs/{run_id}/replan": {
      "post": {
        "operationId": "replan_engineering_run_api_v1_engineering_runs__run_id__replan_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replan Engineering Run Api V1 Engineering Runs  Run Id  Replan Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replan Engineering Run"
      }
    },
    "/api/v1/urls": {
      "post": {
        "operationId": "create_url_api_v1_urls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateURLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/URLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Url"
      }
    },
    "/api/v1/urls/{code}": {
      "delete": {
        "operationId": "disable_url_api_v1_urls__code__delete",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/URLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Url"
      },
      "get": {
        "operationId": "get_url_api_v1_urls__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/URLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Url"
      }
    },
    "/api/v1/urls/{code}/analytics": {
      "get": {
        "operationId": "get_analytics_api_v1_urls__code__analytics_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analytics"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "liveness_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Liveness Health Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness"
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "readiness_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Readiness Health Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness"
      }
    }
  }
}